# echango-api-main/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/v1/countries:
    post:
      operationId: CountriesController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCountryDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
      security:
        - bearer: []
      tags: &ref_4
        - Countries
    get:
      operationId: CountriesController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationCountryResponseDto'
      tags: *ref_4
  /api/v1/countries/{id}:
    get:
      operationId: CountriesController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
      tags: *ref_4
    patch:
      operationId: CountriesController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCountryDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
      security:
        - bearer: []
      tags: *ref_4
    delete:
      operationId: CountriesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_4
  /api/v1/currencies:
    get:
      operationId: CurrenciesController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationCurrencyResponseDto'
      tags: &ref_5
        - Currencies
  /api/v1/currencies/{code}:
    get:
      operationId: CurrenciesController_findByCode_v1
      parameters:
        - name: code
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
      tags: *ref_5
  /api/v1/currencies/search/{query}:
    get:
      operationId: CurrenciesController_searchByCode_v1
      parameters:
        - name: query
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
      tags: *ref_5
...
  /api/v1/faqs:
    post:
      operationId: FaqsController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFaqDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
      security: &ref_7
        - bearer: []
      tags: &ref_8
        - Faqs
    get:
      operationId: FaqsController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationFaqResponseDto'
      security: *ref_7
      tags: *ref_8
  /api/v1/faqs/me:
    get:
      operationId: FaqsController_getMyOfficeFaqs_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationFaqResponseDto'
      security: *ref_7
      tags: *ref_8
    post:
      operationId: FaqsController_createMyOfficeFaq_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMyFaqDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
      security: *ref_7
      tags: *ref_8
  /api/v1/faqs/me/bulk:
    post:
      description: >-
        Allows office owners to create multiple FAQs in a single request. All
        FAQs will be associated with the authenticated user's office.
      operationId: FaqsController_createMyOfficeFaqs_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMyFaqsDto'
      responses:
        '201':
          description: Successfully created FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faq'
        '400':
          description: Invalid input data or validation errors
        '401':
          description: User not authenticated or office not found
      security: *ref_7
      summary: Create multiple FAQs for authenticated user office
      tags: *ref_8
  /api/v1/faqs/me/{id}:
    patch:
      operationId: FaqsController_updateMyOfficeFaq_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFaqDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
      security: *ref_7
      tags: *ref_8
    delete:
      operationId: FaqsController_removeMyOfficeFaq_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '204':
          description: ''
      security: *ref_7
      tags: *ref_8
  /api/v1/faqs/{id}:
    get:
      operationId: FaqsController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
      security: *ref_7
      tags: *ref_8
    patch:
      operationId: FaqsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFaqDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
      security: *ref_7
      tags: *ref_8
    delete:
      operationId: FaqsController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
...
info:
  title: ExchangGo24 API
  description: ' API docs'
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    FileType:
      type: object
      properties:
        id:
          type: string
          example: cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae
        path:
          type: string
          example: https://example.com/path/to/file.jpg
      required:
        - id
        - path
    Role:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: admin
      required:
        - id
        - name
    Status:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: active
      required:
        - id
        - name
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
          example: john.doe@example.com
        provider:
          type: string
          example: email
        socialId:
          type: string
          example: '1234567890'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        photo:
          $ref: '#/components/schemas/FileType'
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - provider
        - socialId
        - firstName
        - lastName
        - photo
        - role
        - status
        - createdAt
        - updatedAt
        - deletedAt
    WorkingHour:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the working hour record
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dayOfWeek:
          type: string
          description: Day of the week
          example: Monday
          enum:
            - Monday
            - Tuesday
            - Wednesday
            - Thursday
            - Friday
            - Saturday
            - Sunday
        isActive:
          type: boolean
          description: Indicates if the day is active for working hours
          example: true
        fromTime:
          type: string
          description: Start time of working hours
          example: '09:00'
          nullable: true
        toTime:
          type: string
          description: End time of working hours
          example: '17:00'
          nullable: true
        hasBreak:
          type: boolean
          description: Indicates if there is a break during this day
          example: false
        breakFromTime:
          type: string
          description: Start time of break
          example: '12:00'
          nullable: true
        breakToTime:
          type: string
          description: End time of break
          example: '13:00'
          nullable: true
        officeId:
          type: string
          description: ID of the associated office
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: Creation timestamp
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          format: date-time
          type: string
          description: Last update timestamp
          example: '2023-01-02T12:00:00Z'
      required:
        - id
        - dayOfWeek
        - isActive
        - fromTime
        - toTime
        - hasBreak
        - breakFromTime
        - breakToTime
        - officeId
        - createdAt
        - updatedAt
    Office:
      type: object
      properties:
        id:
          type: string
        owner:
          $ref: '#/components/schemas/User'
        officeName:
          type: string
        location:
          type: object
        registrationNumber:
          type: string
        currencyExchangeLicenseNumber:
          type: string
        address:
          type: string
        city:
          type: object
        state:
          type: object
        country:
          type: object
        primaryPhoneNumber:
          type: string
        secondaryPhoneNumber:
          type: object
        thirdPhoneNumber:
          type: object
        whatsappNumber:
          type: object
        logo:
          type: object
        mainImage:
          type: object
        slug:
          type: object
        isActive:
          type: boolean
        isVerified:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          type: object
        distanceInKm:
          type: number
        rates:
          type: array
          items:
            type: string
        equivalentValue:
          type: number
        images:
          type: object
        workingHours:
          type: array
          items:
            type: string
        email:
          type: object
        todayWorkingHours:
          $ref: '#/components/schemas/WorkingHour'
      required:
        - id
        - owner
        - officeName
        - location
        - registrationNumber
        - currencyExchangeLicenseNumber
        - address
        - city
        - state
        - country
        - primaryPhoneNumber
        - secondaryPhoneNumber
        - thirdPhoneNumber
        - whatsappNumber
        - logo
        - mainImage
        - slug
        - isActive
        - isVerified
        - isFeatured
        - createdAt
        - updatedAt
        - deletedAt
        - distanceInKm
        - rates
        - equivalentValue
        - images
        - workingHours
        - email
        - todayWorkingHours
...
    Country:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        unicode:
          type: string
        emoji:
          type: string
        alpha2:
          type: string
        dialCode:
          type: string
        region:
          type: string
        capital:
          type: string
        alpha3:
          type: string
      required:
        - id
        - name
        - unicode
        - emoji
        - alpha2
        - dialCode
        - region
        - capital
        - alpha3
...
    CreateCountryDto:
      type: object
      properties:
        name:
          type: string
        unicode:
          type: string
        emoji:
          type: string
        alpha2:
          type: string
        dialCode:
          type: string
        alpha3:
          type: string
        region:
          type: string
        capital:
          type: string
      required:
        - name
        - unicode
        - emoji
        - alpha2
        - dialCode
        - alpha3
        - region
        - capital
    InfinityPaginationCountryResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Country'
        hasNextPage:
          type: boolean
          example: true
      required:
        - data
        - hasNextPage
    UpdateCountryDto:
      type: object
      properties:
        name:
          type: string
        unicode:
          type: string
        emoji:
          type: string
        alpha2:
          type: string
        dialCode:
          type: string
        alpha3:
          type: string
        region:
          type: string
        capital:
          type: string
    Currency:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        namePlural:
          type: string
        symbol:
          type: string
        symbolNative:
          type: string
        decimalDigits:
          type: number
        rounding:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        flag:
          type: string
      required:
        - id
        - code
        - name
        - namePlural
        - symbol
        - symbolNative
        - decimalDigits
        - rounding
        - createdAt
        - updatedAt
        - flag
    InfinityPaginationCurrencyResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Currency'
        hasNextPage:
          type: boolean
          example: true
      required:
        - data
        - hasNextPage
...
    CreateFaqDto:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        officeId:
          type: string
      required:
        - question
        - answer
        - officeId
    Faq:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
        isActive:
          type: boolean
        office:
          $ref: '#/components/schemas/Office'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - question
        - answer
        - isActive
        - office
        - createdAt
        - updatedAt
    InfinityPaginationFaqResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Faq'
        hasNextPage:
          type: boolean
          example: true
      required:
        - data
        - hasNextPage
    CreateMyFaqDto:
      type: object
      properties:
        question:
          type: string
          description: The question for the FAQ
          example: What are your exchange rates?
        answer:
          type: string
          description: The answer for the FAQ
          example: >-
            Our exchange rates are updated daily and are competitive with market
            rates.
      required:
        - question
        - answer
    CreateMyFaqsDto:
      type: object
      properties:
        faqs:
          description: Array of FAQs to create
          example:
            - question: What are your exchange rates?
              answer: >-
                Our exchange rates are updated daily and are competitive with
                market rates.
            - question: What are your working hours?
              answer: We are open Monday to Friday from 9 AM to 6 PM.
          type: array
          items:
            $ref: '#/components/schemas/CreateMyFaqDto'
      required:
        - faqs
    UpdateFaqDto:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        officeId:
          type: string
...